[MAJ] V?rification de l'existance d'un topic de commentaire dans diff?rentes actions.

+++ kernel/framework/content/comments/CommentsService.class.php
@@ -62,7 +62,7 @@ class CommentsService
 			*/
 			else
 			{
-				if (CommentsDAO::get_existed_comments_topic($comments) == 0)
+				if (CommentsDAO::get_existed_comments_topic($comments))
 				{
 					CommentsDAO::create_comments_topic($comments);
 				}
@@ -142,8 +142,10 @@ class CommentsService
 	*/
 	public static function change_visibility(Comments $comments)
 	{
-		// TODO verificate existed module in comments database
-		CommentsDAO::change_visibility($comments);
+		if (CommentsDAO::get_existed_comments_topic($comments))
+		{
+			CommentsDAO::change_visibility($comments);
+		}
 	}
 	
 	/*
@@ -151,8 +153,10 @@ class CommentsService
 	*/
 	public static function delete_comments_module(Comments $comments)
 	{
-		// TODO verificate existed module in comments database
-		CommentsDAO::delete_all_comments_by_module_name($comments);
+		if (CommentsDAO::get_existed_comments_topic($comments))
+		{
+			CommentsDAO::delete_all_comments_by_module_name($comments);
+		}
 	}
 	
 	/*
@@ -160,8 +164,10 @@ class CommentsService
 	*/
 	public static function delete_comments_id_in_module(Comments $comments)
 	{
-		// TODO verificate existed module and id in module in comments database
-		CommentsDAO::delete_comments_id_in_module($comments);
+		if (CommentsDAO::get_existed_comments_topic($comments))
+		{
+			CommentsDAO::delete_comments_id_in_module($comments);
+		}
 	}
 	
 	/*
